% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseuDE2.R
\name{pseuDE2}
\alias{pseuDE2}
\title{pseuDE2}
\usage{
pseuDE2(
  object,
  aggregate_groups_metadata,
  compare_groups_metadata,
  outputDir,
  comp_group1,
  comp_group2 = NULL,
  do_prefilter = T,
  paired = F,
  pct = 0.1,
  order_results = "padj",
  saverds = F,
  return_results = T,
  saveName = NULL,
  cores = 1
)
}
\arguments{
\item{object}{Seurat object. Different layers in the RNA assay will be merged. Only the raw RNA-counts are used in this function.}

\item{aggregate_groups_metadata}{Metadata entry in object. Aggregation will be based on grouping in this metadata.}

\item{compare_groups_metadata}{Metadata entry in object. Should contain at least two different groups for which the DE analysis is performed.}

\item{outputDir}{Path to a folder where results are stored. It will be created if it does not exist.}

\item{comp_group1}{Value present in compare_groups_metadata. This will be the reference group in the comparison.}

\item{comp_group2}{Default = NULL. Value present in compare_groups_metadata. This will be the test group in the comparison. If not set, it automatically takes all cells not in comp_group1. If there are more than 2 groups and this is not set, all other cells will be assigned to a group 'x'.}

\item{do_prefilter}{Default = TRUE. Either TRUE (T) or FALSE (F). When set to TRUE, then only the genes are kept that have more than 2 fragments per million (fpm) in more than half of the samples in either of the two groups.}

\item{paired}{Default = FALSE. Either TRUE (T) or FALSE (F). Whether to perform pseudobulk DE analysis using paired samples.}

\item{pct}{Default = 0.1. Value between 0.0 and 1.0. Ratio of cells in which a gene should be present in either of the two groups in order to be taken into account in the DE analysis.}

\item{order_results}{Default = padj. Should be one of the following strings: 'padj', 'log2FoldChange', 'pvalue', 'baseMean' or 'lfcSE'. Defines how the final results should be ordered. It does not influence the results, only the ordering with which it is saved.}

\item{return_results}{Default = TRUE. Either TRUE (T) or FALSE (F). Whether to return the results (when TRUE) or only save the results in outputDir (when FALSE).}

\item{saveName}{Default = NULL. String. Name to add to the filenames. If NULL, then the output files will be named 'log2fc.csv' and, if saveRDS == TRUE, 'DESeq2object.RDS'. For example, if saveName = 'sample1', then the filenames will become 'log2fc_sample1.csv' and 'DESeq2object_sample1.RDS'.}

\item{cores}{Default = 1. Positive integer value. How many cores to be used when running DESeq2. When value is > 1, then BiocParallel is used for parallelization.}

\item{saveRDS}{Default = FALSE. Either TRUE (T) or FALSE (F). Whether to save the DESeq2 object in outputDir.}
}
\value{
When return_results = TRUE: List containing two objects:
\itemize{
\item A matrix with the DESeq2 lfcShrink output and the normalized counts
\item The DESeq2 output object.
}
}
\description{
Perform differential expression analysis using pseudobulk.
It requires a Seurat object as main input.
The different metadata entries in the Seurat object can be used to define how the data needs to be aggregated and which comparison should be made.
A positive log2FC means an up-regulation in comp_group1 relative to comp_group2 (or all other cells when comp_group2 is not defined).
}
\details{
Normal differential expression analysis (e.g. using Seurat::FindMarkers) treats all cells as independent measurements.
Strictly speaking this is not fair as individual cells from the same sample are not independent.
This also explains the inflated p-values one typically obtains when performing differential expression analysis with single cell data.
A common method to make the measurements independent is by summing the counts of individual cells per sample per group (e.g. cell type).
This creates a single independent 'measurement' where the measurement is now representing a bulk RNA experiment per sample per group (called pseudobulk).
Now we can compare the independent measurements between conditions (e.g. before and after treatment).
For a proper statistical test this requires at least 5 independent samples per condition (thus at least 5 single cell experiments per condition).
Pseudobulk also allows for performing a paired differential analysis.
This can be done when the same samples are sequenced between conditions.
The between sample variability can then be accounted for when performing the differential expression analysis between conditions.

Design equations:
\itemize{
\item Unpaired analysis: '~ + com_group'
\item Paired analysis: '~ + pair_group + com_group'
}

EXAMPLE

We want to compare conditionA with conditionB in T-cells using 5 patients.
There are two approaches we can take.

Approach 1:

We create a metadata entry where we concatenate the metadata that defines the conditions and the cell types.
We would then get, for each cell, something like cellType_condition (e.g. Tcell_conditionA; Tcell_conditionB, neutrophil_conditionA, etc.).
This allows us to tell the function to compare conditionA with conditionB only for the T-cells.
This new metadata entry is going to be our choice for 'compare_groups_metadata' together with comp_group1 = 'Tcell_conditionA' and comp_group2 = 'Tcell_conditionB'.

Approach 2:

We subset our Seurat object to only include T-cells, for example with \code{subset(object, subset = cellTypes == 'Tcells')}.
Now we can make the comparison between conditionA and conditionB directly without adding a new metadata entry.
The entry for 'compare_groups_metadata' is now 'Condition' with comp_group1 = 'conditionA' and comp_group2 = 'conditionB'.

For either approach, the metadata that defines the patients is going to be our entry for 'aggregate_groups_metadata'.
Since the same patients are present in both conditionA and conditionB, when can perform a paired comparison by setting paired = T. This corrects for any biological variability wihin the patients between the conditions.
When paired = TRUE, the function creates a table with two columns called 'com_group' and 'pair_group':

\tabular{llrr}{
\tab \tab \strong{com_group} \tab \strong{pair_group} \cr
pat1-Tcell_conditionA \tab \tab Tcell_conditionA \tab pat1 \cr
pat1-Tcell_conditionB \tab \tab Tcell_conditionB \tab pat1 \cr
pat2-Tcell_conditionA \tab \tab Tcell_conditionA \tab pat2 \cr
pat2-Tcell_conditionB \tab \tab Tcell_conditionB \tab pat2 \cr
pat1-Tcell_conditionA \tab \tab Tcell_conditionA \tab pat3 \cr
pat1-Tcell_conditionB \tab \tab Tcell_conditionB \tab pat3 \cr
pat1-Tcell_conditionA \tab \tab Tcell_conditionA \tab pat4 \cr
pat1-Tcell_conditionB \tab \tab Tcell_conditionB \tab pat4 \cr
pat1-Tcell_conditionA \tab \tab Tcell_conditionA \tab pat5 \cr
pat1-Tcell_conditionB \tab \tab Tcell_conditionB \tab pat5 \cr
}

In case paired = FALSE, the 'pair_group' column is not created and the comparison is only made based on the 'com_group' column.
}
\examples{
\dontrun{
pseude_res <- pseuDE2(sr,
aggregate_groups_metadata = 'Patients',
compare_group_metadata = 'cellType_condition',
outputDir = 'output/',
comp_group1 = 'Tcell_conditionA',
comp_group2 = 'Tcell_conditionB',
paired = T,
cores = 3)

res_counts <- pseude_res[[1]] # Table with log2FC, p-values and normalized counts.
de_out <- pseude_res[[2]] # DESeq2 object

# CREATE VOLCANOPLOT ----
EnhancedVolcano(res_counts,
                rownames(res_counts),
                x ="Log2FoldChange",
                y ="padj",
                drawConnectors = TRUE,
                cutoffLineWidth = 0.3,
                widthConnectors = 0.2,
                max.overlaps = 60,
                xlab = bquote(~Log[2]~ "fold change"),
                title = paste0("Pseudobulk DEG"),
                pCutoff = 0.05,
                FCcutoff = 0.499,
                gridlines.major = FALSE,
                gridlines.minor = FALSE
)

DimPlot(sr,
        reduction = 'umap',
        group.by = 'Patients',
        raster = F) +
    DimPlot(sr,
            reduction = 'umap',
            group.by = 'cellType_condition',
            raster = F) +
    FeaturePlot(sr,
                features = head(rownames(deg_table), 5),
                raster = F,
                order = T)

}

}
